requirements.txt
# core
tqdm==4.66.4

# retrieval/indexing (later steps)
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
torch==2.3.1
transformers==4.43.3
